feat: change default Elasticsearch API version to 2.4
Newer versions of `elasticsearch-js` ship only with support for the highest minor
version of old major elasticsearch versions. We currently use 2.3 in
production for Mapzen Search, but as mentioned in the changelog[1],
using API version 2.4 will work just fine. As defined in the rules of
semver, only new API calls could be added, in 2.4, so anything that
works in 2.3 will still be fine.

Changing this version allows us to upgrade to newer versions of
`elasticsearch-js` which support the latest versions of Elasticsearch 5,
making it easier to upgrade.

[1] https://www.elastic.co/guide/en/elasticsearch/client/javascript-api/current/changelog.html#_13_0_0_apr_24_2017