feat(generate): Add custom and default config helper functions
This adds two new functions to the `pelias-config` module:

`generateDefault`: this function always returns the default config, with
the `get` function added. While we currently expose a `defaults` object
in the module, this won't have the very useful get function added in
https://github.com/pelias/config/pull/101

`generateCustom`: this function returns a config object with custom
changes applied. It takes two parameters: `custom_config, which is an
object just like what might be found in a users `pelias.json`, and
`deep`, which determines if a deep or shallow merge is performed.

Like, `generate`, deep defaults to true. And like `generateDefault`, a
proper `get` function is added to all returned objects.

Both of these functions can be very useful in unit tests. Without them,
it's a lot of work to create valid `pelias-config` objects.