feat(elasticsearch): support `@elastic/elasticsearch` client
BREAKING: Support new elasticsearch configuration format.

As part of updating to elasticsearch 8, I'm updating pelias from the
deprecated `elastic` package to it's successor `@elastic/elasticsearch`
which supports both es7 and 8 (and some older versions too, but I think we
only care about 7+).

Note that this is a backwards incompatible change - old pelias code
using the `elasticsearch` package will not be able to use these new
configs.

apiVersion is determined by the client version, so it's no longer
necessary.

`keepAlive` is true by default, you need to set `agent: false` to
disable it:

    const client = new Client({
      node: 'http://localhost:9200',
      // Disable agent and keep-alive
      agent: false
    })

`hosts` is now `nodes` (or `node`) and has a different format.

Deleted parameters that aren't used in the new client:

 - `env`: I'm not sure what it's for. I suspect it's a pelias thing, not
   an elasticsearch thing, which seems like a conflation of the
   responsibilities of 'esclient config' vs 'pelias env mgmt'. Maybe I'm
   confused though.

 - `log`: The new elasticsearch client emits events instead of logging.